Move everything related to Apache HTTPClient to new ApacheBaseApi (#470)
