+++ opacclient/libopac/src/main/java/de/geeksfactory/opacclient/apis/ApacheSearchOnlyApi.java
@@ -12,7 +12,7 @@ import de.geeksfactory.opacclient.objects.DetailedItem;
  * For search-only API implementations, i.e. without account support, some abstract method
  * implementations are provided.
  */
-public abstract class SearchOnlyApi extends BaseApi {
+public abstract class ApacheSearchOnlyApi extends ApacheBaseApi {
 
     @Override
     public final ReservationResult reservation(DetailedItem item, Account account,
