+++ opacclient/meaningdetector/src/main/java/de/geeksfactory/opacclient/meanings/GetSearchFieldsCallable.java
@@ -10,7 +10,7 @@ import java.util.Set;
 import java.util.concurrent.Callable;
 
 import de.geeksfactory.opacclient.OpacApiFactory;
-import de.geeksfactory.opacclient.apis.BaseApi;
+import de.geeksfactory.opacclient.apis.ApacheBaseApi;
 import de.geeksfactory.opacclient.apis.OpacApi;
 import de.geeksfactory.opacclient.apis.OpacApi.OpacErrorException;
 import de.geeksfactory.opacclient.objects.Library;
@@ -27,8 +27,8 @@ public class GetSearchFieldsCallable implements Callable<Map<String, List<Search
     public Map<String, List<SearchField>> call() {
         OpacApi api = OpacApiFactory.create(lib, "OpacApp/Test");
 
-        if (api instanceof BaseApi) {
-            ((BaseApi) api).setHttpLoggingEnabled(false);
+        if (api instanceof ApacheBaseApi) {
+            ((ApacheBaseApi) api).setHttpLoggingEnabled(false);
         }
 
         Set<String> langs = null;
